# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SDMAS API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/attendance/section/{section_id}": {
      "get": {
        "tags": [
          "attendance"
        ],
        "summary": "Get Section Attendance",
        "operationId": "get_section_attendance_attendance_section__section_id__get",
        "parameters": [
          {
            "name": "section_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Section Id"
            }
          },
          {
            "name": "attendance_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Date to query",
              "title": "Attendance Date"
            },
            "description": "Date to query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AttendanceRecordResponse"
                  },
                  "title": "Response Get Section Attendance Attendance Section  Section Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AttendanceRecordResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "student_id": {
            "type": "integer",
            "title": "Student Id"
          },
          "academic_year_id": {
            "type": "integer",
            "title": "Academic Year Id"
          },
          "class_id": {
            "type": "integer",
            "title": "Class Id"
          },
          "section_id": {
            "type": "integer",
            "title": "Section Id"
          },
          "attendance_date": {
            "type": "string",
            "title": "Attendance Date"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "recorded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Recorded At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "student_id",
          "academic_year_id",
          "class_id",
          "section_id",
          "attendance_date",
          "status",
          "recorded_at",
          "updated_at"
        ],
        "title": "AttendanceRecordResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}